Adding limited CPS support for fn expressions.
In addition to rich language of atomic types (e.g. numbers, booleans,
keywords, etc.) accepted by the cps macro, it now supports fn
expressions that have a single trivial expression (in the Danvy first
order one pass CPS algorithm sense) as the body.

Extensive testing has been performed in the form of CPSing the identity
function and posting a screenshot of the results to Twitter.