Adding partial fix for the arity clash problem.
This commit partially fixes the arity clash problem described in
the previous commit. When the arity of a CPS version of one fn
expression overload matches the arity of one of the original
overloads, a merged overload is generated such that it uses the CPS
version's parameter list, and the body of the is modified to become
an if expression that checks the first argument for the semi-unique
metadata tag attached to each continuation, and if present executes
the CPS version of the code. If the tag isn't present, the parameters
are rebound to the names of those used for the non-CPS version and
the compatibility call to the overload is executed.

For example, consider the following (simplified) transformation:

  => (cps (fn ([x] x) ([x y] (x y))))
  (fn my-fn
    ([x] (my-fn (with-meta identity {:tag true}) x))
    ([k x]
     (if (:tag (meta k))
         (k x)
         (let [y x
               x k]
           (my-fn (with-meta identity {:tag true}) x y))))
    ([k x y] (x k y)))

While these changes solve the problem of CPS vs. non-CPS calls for
fixed-arity overloading, it doesn't solve it for variadic functions.